微調整

+++ BlackJack/src/main/java/dao/UserDao.java
@@ -29,13 +29,14 @@ public class UserDao extends BaseDao {
 		if (password.equals(password2)) {
 
 			try {
-				// SQL文
-				String sql = "INSERT INTO users(user_id, nickname, password) VALUES (?, ?, ?)";
-				ps = con.prepareStatement(sql);
-				ps.setString(1, userId);
-				ps.setString(2, nickname);
-				ps.setString(3, password);
-				ps.executeUpdate();
+				if (con != null) {
+					String sql = "INSERT INTO users(user_id, nickname, password) VALUES (?, ?, ?)";
+					ps = con.prepareStatement(sql);
+					ps.setString(1, userId);
+					ps.setString(2, nickname);
+					ps.setString(3, password);
+					ps.executeUpdate();
+				}
 
 			} catch (SQLIntegrityConstraintViolationException e) {
 				e.printStackTrace();
@@ -56,20 +57,21 @@ public class UserDao extends BaseDao {
 		User user = new User();
 
 		try {
-			// SQL文
-			String sql = "SELECT * FROM users WHERE (user_id = ?) && (password = ?)";
-			ps = con.prepareStatement(sql);
-			ps.setString(1, userId);
-			ps.setString(2, password);
-			rs = ps.executeQuery();
+			if (con != null) {
+				String sql = "SELECT * FROM users WHERE (user_id = ?) && (password = ?)";
+				ps = con.prepareStatement(sql);
+				ps.setString(1, userId);
+				ps.setString(2, password);
+				rs = ps.executeQuery();
 
-			if (rs.next()) {
-				// ログイン情報が正しい場合の処理
-				user.setUserId(rs.getString("user_id"));
-				user.setNickname(rs.getString("nickname"));
-			} else {
-				// ログイン情報が誤っている場合の処理
-				message = "ログインに失敗しました。";
+				if (rs.next()) {
+					// ログイン情報が正しい場合の処理
+					user.setUserId(rs.getString("user_id"));
+					user.setNickname(rs.getString("nickname"));
+				} else {
+					// ログイン情報が誤っている場合の処理
+					message = "ログインに失敗しました。";
+				}
 			}
 
 		} catch (SQLException e) {
@@ -85,11 +87,12 @@ public class UserDao extends BaseDao {
 	public void doDelete(User loginUser) {
 
 		try {
-			// SQL文
-			String sql = "DELETE FROM users WHERE user_id = ?";
-			ps = con.prepareStatement(sql);
-			ps.setString(1, loginUser.getUserId());
-			ps.executeUpdate();
+			if (con != null) {
+				String sql = "DELETE FROM users WHERE user_id = ?";
+				ps = con.prepareStatement(sql);
+				ps.setString(1, loginUser.getUserId());
+				ps.executeUpdate();
+			}
 
 		} catch (SQLException e) {
 			e.printStackTrace();
@@ -101,20 +104,22 @@ public class UserDao extends BaseDao {
 
 	/**
 	 * DBのユーザ情報のユーザIDとニックネームを変更する。
+	 * @return 変更後のユーザIDとニックネームを有するUserオブジェクト
 	 */
 	public User editIdName(String userId, String nickname, String sessionUserId) {
 
 		User loginUser = new User();
 
 		try {
-			// SQL文
-			String sql = "UPDATE users SET user_id = ?, nickname = ? WHERE user_id = ?";
-			ps = con.prepareStatement(sql);
-			ps.setString(1, userId);
-			ps.setString(2, nickname);
-			ps.setString(3, sessionUserId);
-			ps.executeUpdate();
-			message = "ユーザID、ニックネームを変更しました。";
+			if (con != null) {
+				String sql = "UPDATE users SET user_id = ?, nickname = ? WHERE user_id = ?";
+				ps = con.prepareStatement(sql);
+				ps.setString(1, userId);
+				ps.setString(2, nickname);
+				ps.setString(3, sessionUserId);
+				ps.executeUpdate();
+				message = "ユーザID、ニックネームを変更しました。";
+			}
 
 		} catch (SQLIntegrityConstraintViolationException e) {
 			e.printStackTrace();
@@ -140,21 +145,22 @@ public class UserDao extends BaseDao {
 		if (newPassword.equals(newPassword2)) {
 
 			try {
-				// SQL文
-				String sql = "UPDATE users SET password = ? WHERE (user_id = ?) && (password = ?)";
-				ps = con.prepareStatement(sql);
-				ps.setString(1, newPassword);
-				ps.setString(2, sessionUserId);
-				ps.setString(3, oldPassword);
+				if (con != null) {
+					String sql = "UPDATE users SET password = ? WHERE (user_id = ?) && (password = ?)";
+					ps = con.prepareStatement(sql);
+					ps.setString(1, newPassword);
+					ps.setString(2, sessionUserId);
+					ps.setString(3, oldPassword);
 
-				int changedRows = ps.executeUpdate();
+					int changedRows = ps.executeUpdate();
 
-				if (changedRows == 0) {
-					message = "古いパスワードが間違っています。";
-				} else {
-					message = "パスワードを変更しました。";
+					if (changedRows == 0) {
+						message = "古いパスワードが間違っています。";
+					} else {
+						message = "パスワードを変更しました。";
+					}
 				}
-				
+
 			} catch (SQLException e) {
 				e.printStackTrace();
 				message = "SQL実行中に例外が発生しました";
@@ -166,42 +172,52 @@ public class UserDao extends BaseDao {
 		}
 	}
 
+	/**
+	 * usersテーブルから総プレイ人口を求める。
+	 */
 	public int getPopulation() {
 
 		int population = 0;
 
 		try {
-			// SQL文
-			String sql = "SELECT * FROM users";
-			ps = con.prepareStatement(sql);
-			rs = ps.executeQuery();
+			if (con != null) {
+				String sql = "SELECT count(*) FROM users";
+				ps = con.prepareStatement(sql);
+				rs = ps.executeQuery();
 
-			while (rs.next()) {
-				population += 1;
+				rs.next();
+				population = rs.getInt("count(*)");
 			}
+
 		} catch (SQLException e) {
 			e.printStackTrace();
 			message = "SQL実行中に例外が発生しました";
-		} 
+		}
 		return population;
 	}
-	
+
+	/**
+	 * usersテーブルのwin_rateカラムをもとに勝率ランキングトップ5を生成する。
+	 * @return ニックネームと勝率を有するUserオブジェクトのリスト
+	 */
 	public List<User> getRankingList() {
 
 		List<User> rankingList = new ArrayList<User>();
 
 		try {
-			// SQL文
-			String sql = "SELECT * FROM users ORDER BY win_rate DESC LIMIT 5";
-			ps = con.prepareStatement(sql);
-			rs = ps.executeQuery();
+			if (con != null) {
+				String sql = "SELECT * FROM users ORDER BY win_rate DESC LIMIT 5";
+				ps = con.prepareStatement(sql);
+				rs = ps.executeQuery();
 
-			while (rs.next()) {
-				User user = new User();
-				user.setNickname(rs.getString("nickname"));
-				user.setWinRate(rs.getFloat("win_rate"));
-				rankingList.add(user);
+				while (rs.next()) {
+					User user = new User();
+					user.setNickname(rs.getString("nickname"));
+					user.setWinRate(rs.getFloat("win_rate"));
+					rankingList.add(user);
+				}
 			}
+
 		} catch (SQLException e) {
 			e.printStackTrace();
 			message = "SQL実行中に例外が発生しました";
@@ -210,25 +226,30 @@ public class UserDao extends BaseDao {
 		}
 		return rankingList;
 	}
-	
+
+	/**
+	 * usersテーブルからログイン中ユーザの情報を取得する。
+	 * @return 勝利回数、敗北回数、引き分け回数、勝率を有するUserオブジェクト
+	 */
 	public User getUserInfo(User loginUser) {
-		
+
 		User user = new User();
 		String userId = loginUser.getUserId();
-		
+
 		try {
-			//SQL文
-			String sql = "SELECT * FROM users WHERE user_id = ?";
-			ps = con.prepareStatement(sql);
-			ps.setString(1, userId);
-			rs = ps.executeQuery();
-			
-			rs.next();
-			user.setWin(rs.getInt("win"));
-			user.setLose(rs.getInt("lose"));
-			user.setDraw(rs.getInt("draw"));
-			user.setWinRate(rs.getFloat("win_rate"));
-			
+			if (con != null) {
+				String sql = "SELECT * FROM users WHERE user_id = ?";
+				ps = con.prepareStatement(sql);
+				ps.setString(1, userId);
+				rs = ps.executeQuery();
+
+				rs.next();
+				user.setWin(rs.getInt("win"));
+				user.setLose(rs.getInt("lose"));
+				user.setDraw(rs.getInt("draw"));
+				user.setWinRate(rs.getFloat("win_rate"));
+			}
+
 		} catch (SQLException e) {
 			e.printStackTrace();
 			message = "SQL実行中に例外が発生しました";
@@ -237,27 +258,30 @@ public class UserDao extends BaseDao {
 		}
 		return user;
 	}
-	
+
+	/**
+	 * usersテーブルのwin,lose,drawカラムのいずれかに1を加算し、win_rateカラムを更新する。
+	 */
 	public void updateResult(User loginUser, History history) {
-		
+
 		String userId = loginUser.getUserId();
 		String strResult = history.getStrResult();
-		
-		try {
-			// SQL文(win,lose,drawカラムのいずれかを更新)
-			String sql = String.format("UPDATE users SET %s = %s + 1 WHERE user_id = ?", strResult, strResult);
-			ps = con.prepareStatement(sql);
-			ps.setString(1, userId);
-			System.out.println(ps);
-			ps.executeUpdate();
 
-			// SQL文(win_rateカラムを更新)
-			sql = "UPDATE users SET win_rate = 100 * win / (win + lose + draw) WHERE user_id = ?";
-			ps = con.prepareStatement(sql);
-			ps.setString(1, userId);
-			System.out.println(ps);
-			ps.executeUpdate();
-			
+		try {
+			if (con != null) {
+				// win,lose,drawカラムのいずれかに1を加算
+				String sql = String.format("UPDATE users SET %s = %s + 1 WHERE user_id = ?", strResult, strResult);
+				ps = con.prepareStatement(sql);
+				ps.setString(1, userId);
+				ps.executeUpdate();
+
+				// win_rateカラムを更新
+				sql = "UPDATE users SET win_rate = 100 * win / (win + lose + draw) WHERE user_id = ?";
+				ps = con.prepareStatement(sql);
+				ps.setString(1, userId);
+				ps.executeUpdate();
+			}
+
 		} catch (SQLException e) {
 			e.printStackTrace();
 			message = "SQL実行中に例外が発生しました";
